+++ web/src/lib.rs
@@ -14,6 +14,8 @@
 #[cfg(target_family = "wasm")]
 mod c_alloc;
 #[cfg(target_family = "wasm")]
+mod clipboard;
+#[cfg(target_family = "wasm")]
 mod clock;
 #[cfg(target_family = "wasm")]
 mod download;